Move clockwise() to geometry.ts
